Store temporary files in `SRC_DIR`
Write all temporary files in `SRC_DIR`. This way they are contained with
the conda-build created directories (and cleaned up afterwards). Also
for Linux builds, this has the advantage of writing to a storage
location that resides outside the container. This can make their
contents easier to inspect. It should also minimize storage used by the
container itself in Linux builds, which in turn should fix the CUDA 11.8
Linux x86_64 CI builds that are running out of space.