Specify runfile's `--tmpdir` within `$SRC_DIR/tmp`
Ensure when the runfile uses a temporary directory for any extraction
work it is doing, that the work occurs with `$SRC_DIR/tmp`, which is a
conda-build managed location (instead of a system one). Thus it lives
outside containers on Linux and gets cleaned up after builds. This
should avoid running out of memory when containers are used.